med-hc-whatsapp-mvp-v2_2-patched/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfminer.six==20240706
httpx==0.27.0
pydantic==2.8.2
apscheduler==3.10.4
python-dotenv==1.0.1
